addedd disparate impact fairness metric in three metric tutorial
