+++ README.md
@@ -79,7 +79,7 @@ We have prepared interactive demos, where you can play around without having to 
 ### Fairness of Machine Learning
 |Name| Notebook           | Task  | Example                       |
 |:---------------------------------:|:-------------:|:-----:|:------------:|
-Demographic parity <br> Equal opportunity <br> Equalised odds | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sony/nnabla-examples/blob/master/interactive-demos/three_metrics.ipynb) | [**Metrics tutorial**] <br> Dataset/Model Bias Check |<a href="url"><img src="https://github.com/sony/nnabla-examples/raw/master/responsible_ai/three_metrics/images/three_metrics.png" align="center" height="90" ></a>|
+Demographic parity <br> Disparate Impact <br> Equal opportunity <br> Equalised odds | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sony/nnabla-examples/blob/master/interactive-demos/fairness_metrics.ipynb) | [**Metrics tutorial**] <br> Dataset/Model Bias Check |<a href="url"><img src="https://github.com/sony/nnabla-examples/raw/master/responsible_ai/fairness_metrics/images/fairness_metrics.PNG" align="center" height="90" ></a>|
 Reweighing | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sony/nnabla-examples/blob/master/interactive-demos/gender_bias_mitigation_german_cc.ipynb) |  [**Pre-processing tutorial**] <br> Dataset/Model Bias Check and Mitigation by Reweighing |<a href="url"><img src="https://github.com/sony/nnabla-examples/raw/master/responsible_ai/gender_bias_mitigation/images/fairness.png" align="center" height="90" ></a>|
 GAN Data Debiasing | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sony/nnabla-examples/blob/master/interactive-demos/gender_bias_mitigation_german_cc.ipynb) |  [**Pre-processing tutorial**] <br> Dataset/Model Bias Check and Mitigation by GAN |<a href="url"><img src="https://github.com/sony/nnabla-examples/raw/master/responsible_ai/gan_data_debiased/images/gan_data_debiasing_workflow.png" align="center" height="90" ></a>|
 Prejudice Remover Regularizer | [![Open In Colab](https://colab.research.google.com/assets/colab-badge.svg)](https://colab.research.google.com/github/sony/nnabla-examples/blob/master/interactive-demos/prejudice_remover_regularizer.ipynb) |  [**In-processing tutorial**] <br> Model Bias Check and Mitigation by Prejudice Removal Technique  |<a href="url"><img src="https://github.com/sony/nnabla-examples/blob/master/responsible_ai/prejudice_remover_regularizer/images/Prejudice_Remover_Regularizer_Equation.png" align="center" height="90" ></a>|
